Fix for ray-gcs server errors on multiple restarts
If we cache the tmp directory, it turns out that ray looks for the last
time we initialize the server with the same port and tries to re-connect.
It does this by looking for the cached `/tmp/ray/ray_current_cluster`
file, however, the cluster is no longer running and so it fails.

This new implementation fixes this by forcing ray to start a new
cluster, via `ray.init(address="local")` and then connecting to the
server via `ray.init(address="auto")`. The first call to
`ray.init(address="auto")` raises a `ConnectionError` and proceeds to
force-start a new ray cluster instance, followed by a second call to
`ray.init(address="auto")` which successfully connects to the server.