Shared memory transport for gRPC
This PR adds a new transport for gRPC that uses shared memory for
communication between client and server. This transport is intended
to be used for inference workloads where the client and server are on
the same machine. The transport is implemented using the shared memory
primitives in the multiprocessing module.

 - Adds new `shm.py` module in `nos/common` with shared memory primitives
 - Adds tests for the new transport with new `NOS_SHM_ENABLED` env var
 - Updates to volume mounts for all docker-compose files
 - Updated Dockerfile with `NOS_SHM_ENABLED=1` enabled by default
 - Updates to `nos_service.proto` to register/unregister `SHM` transport