Twine upload on release (#350)
Attempting to sidestep a hairy issue where the `.whl` distribution name
doesn't quite match that supplied to the `pyproject.toml` via python
version import. Specifically, it replaces `-` with `.` and appends a
trailing 0, neither of which are found in the $NOS_VERSION. Workaround
is to grep for the base version with some name mangling. Also setting up
twine permissions to pull uname and password as github secrets so we can
keep these out of the repo. `twine` needs to be a dependency alongside
`build`. Moved these to `requirements.dev.txt`, need to confirm they
still get installed with `develop-cpu`.

<!-- For example: "Closes #1234" -->

## Checks

- [x] `make lint`: I've run `make lint` to lint the changes in this PR.
- [x] `make test`: I've made sure the tests (`make test-cpu` or `make
test`) are passing.
- Additional tests:
   - [ ] Benchmark tests (when contributing new models)
   - [ ] GPU/HW tests