Support inferred and non-dict `return_annotations`
In order to support custom models and automatic model registry, we need
to be able to inspect the function signatures of the model methods and
infer their return types. This PR addresses this by removing the need to
only return dictionaries from the model methods. Instead, we now support
returning any type of object.