New `nos serve` CLI for local serving of custom models (#438)
## Summary

- **Custom model serving via `nos-serve`**: In order to support custom
models externally to NOS, we introduce a model catalog that can be
dynamically imported and built with a custom runtime image (derived from
the base gpu image). For more details, see `nos/cli/serve.py`.
- **Support inferred and non-dict return_annotations**: In order to
support custom models and automatic model registry, we need
to be able to inspect the function signatures of the model methods and
infer their return types. This PR addresses this by removing the need to
only return dictionaries from the model methods. Instead, we now support
returning any type of object.
- Remove internal library patching for open-source release

<!-- Please give a short summary of the change and the problem this
solves. -->

## Related issues

<!-- For example: "Closes #1234" -->

## Checks

- [ ] `make lint`: I've run `make lint` to lint the changes in this PR.
- [ ] `make test`: I've made sure the tests (`make test-cpu` or `make
test`) are passing.
- Additional tests:
   - [ ] Benchmark tests (when contributing new models)
   - [ ] GPU/HW tests