Support file-uploads with new HTTP endpoint
 This PR supports file-uploads for various models (object detection,
 whisper etc) where the user is allowed to upload a file and get the
inference results. The file-upload is supported by the new HTTP endpoint
`/infer_file` POST request. The request body should contain the file
data and the request headers should contain the `model_id` and the
`method` (optional), with the `file`. A few examples have been added to
the `README` along with HTTP client-side tests to verify the functionality.