Support model streaming responses with new `Stream` gRPC service
This PR adds a new `Stream` gRPC service that allows for streaming
responses from the server. This is useful for models that produce
a stream of outputs such as text, video frames or audio chunks. The
`Stream` service is implemented as a uni-directional output stream
with a streaming response.

New models:
 - `llama2_chat` - a simple chatbot model that uses the `transformers`
   `AutoModelForCausalLM` with a `chat` submethod. New models include:
   `meta-llama/Llama-2-7b-chat-hf`, `HuggingFaceH4/zephyr-7b-beta`,
   `HuggingFaceH4/tiny-random-LlamaForCausalLM` (for testing purposes).

Other changes:
 - refactored names for all exceptions without `Nos` prefix