Support model streaming responses with new `Stream` gRPC service (#469)
This PR adds a new `Stream` gRPC service that allows for streaming
responses from the server. This is useful for models that produce a
stream of outputs such as text, video frames or audio chunks. The
`Stream` service is implemented as a uni-directional output stream with
a streaming response.

New models:
- `llama2_chat` - a simple chatbot model that uses the `transformers`
`AutoModelForCausalLM` with a `chat` submethod. New models include:
`meta-llama/Llama-2-7b-chat-hf`, `HuggingFaceH4/zephyr-7b-beta`,
`HuggingFaceH4/tiny-random-LlamaForCausalLM` (for testing purposes).

Other changes:
 - refactored names for all exceptions without `Nos` prefix

<!-- Thank you for your contribution! Please review
https://github.com/autonomi-ai/nos/blob/main/docs/CONTRIBUTING.md before
opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR.
If you don't have the access to it, we will shortly find a reviewer and
assign them to your PR. -->

## Summary

<!-- Please give a short summary of the change and the problem this
solves. -->

## Related issues

<!-- For example: "Closes #1234" -->

## Checks

- [ ] `make lint`: I've run `make lint` to lint the changes in this PR.
- [ ] `make test`: I've made sure the tests (`make test-cpu` or `make
test`) are passing.
- Additional tests:
   - [ ] Benchmark tests (when contributing new models)
   - [ ] GPU/HW tests