Support async gRPC handling with scaled-replicas (#494)
## Summary

This PR adds support for async gRPC handling with scaled
replicas. Individual requests are dispatched to one of the many
actor pools and the response is fetched by the client asynchronously.

We also introduce ActorPools with task throttling to make sure we can
scale
and await on tasks using the async gRPC service. 

Other additions / fixes:
- Support loading models with fixed number of replicas on init
- Improved `ModelResource` handling for gpu/cpu memory resources
- All `noop` methods are now CPU only models
- Refactored grpc server tests to re-use serve fixture
- Fixed HTTP client API implementation to use normalized model-ids

## Related issues

<!-- For example: "Closes #1234" -->

## Checks

- [ ] `make lint`: I've run `make lint` to lint the changes in this PR.
- [ ] `make test`: I've made sure the tests (`make test-cpu` or `make
test`) are passing.
- Additional tests:
   - [ ] Benchmark tests (when contributing new models)
   - [ ] GPU/HW tests