Override `ModelDeploymentSpec` of registered models
This PR allows users to override the deployment spec of registered models.
For example, registered models and their deployment spec can be overridden as follows:

```yaml
TinyLlama/TinyLlama-1.1B-Chat-v1.0:
    runtime_env: custom-gpu
    deployment:
      resources:
        device: auto
        device_memory: 3Gi
```

In this example, the deployment resources for the model
`TinyLlama/TinyLlama-1.1B-Chat-v1.0` will be overridden to use
the `custom-gpu` runtime environment and `3Gi` of device memory.